+++ taf/yubikey.py
@@ -95,12 +95,11 @@ def raise_yubikey_err(msg: Optional[str] = None) -> Callable:
 
 
 @contextmanager
-def _yk_piv_ctrl(serial=None, pub_key_pem=None):
+def _yk_piv_ctrl(serial=None):
     """Context manager to open connection and instantiate Piv Session.
 
     Args:
-        - pub_key_pem(str): Match Yubikey's public key (PEM) if multiple keys
-                            are inserted
+        - serial (str): Match Yubikey's serial multiple keys are inserted
 
     Returns:
         - ykman.piv.PivSession
@@ -110,35 +109,13 @@ def _yk_piv_ctrl(serial=None, pub_key_pem=None):
     """
     # If pub_key_pem is given, iterate all devices, read x509 certs and try to match
     # public keys.
-    if pub_key_pem is not None:
-        for dev, info in list_all_devices():
-            # Connect to a YubiKey over a SmartCardConnection, which is needed for PIV.
+    for dev, info in list_all_devices():
+        if serial is None or info.serial == serial:
             with dev.open_connection(SmartCardConnection) as connection:
                 session = PivSession(connection)
-                device_pub_key_pem = (
-                    session.get_certificate(SLOT.SIGNATURE)
-                    .public_key()
-                    .public_bytes(
-                        encoding=serialization.Encoding.PEM,
-                        format=serialization.PublicFormat.SubjectPublicKeyInfo,
-                    )
-                    .decode("utf-8")
-                )
-                # Tries to match without last newline char
-                if (
-                    device_pub_key_pem == pub_key_pem
-                    or device_pub_key_pem[:-1] == pub_key_pem
-                ):
-                    break
                 yield session, info.serial
-    else:
-        for dev, info in list_all_devices():
-            if serial is None or info.serial == serial:
-                with dev.open_connection(SmartCardConnection) as connection:
-                    session = PivSession(connection)
-                    yield session, info.serial
-            else:
-                pass
+        else:
+            pass
 
 
 def is_inserted():
@@ -178,7 +155,7 @@ def is_valid_pin(pin):
 
 
 @raise_yubikey_err("Cannot get serial number.")
-def get_serial_num(pub_key_pem=None):
+def get_serial_num():
     """Get Yubikey serial number.
 
     Args:
@@ -191,12 +168,12 @@ def get_serial_num(pub_key_pem=None):
     Raises:
         - YubikeyError
     """
-    with _yk_piv_ctrl(pub_key_pem=pub_key_pem) as (_, serial):
+    with _yk_piv_ctrl() as (_, serial):
         return serial
 
 
 @raise_yubikey_err("Cannot export x509 certificate.")
-def export_piv_x509(cert_format=serialization.Encoding.PEM, pub_key_pem=None):
+def export_piv_x509(cert_format=serialization.Encoding.PEM):
     """Exports YubiKey's piv slot x509.
 
     Args:
@@ -210,13 +187,13 @@ def export_piv_x509(cert_format=serialization.Encoding.PEM, pub_key_pem=None):
     Raises:
         - YubikeyError
     """
-    with _yk_piv_ctrl(pub_key_pem=pub_key_pem) as (ctrl, _):
+    with _yk_piv_ctrl() as (ctrl, _):
         x509 = ctrl.get_certificate(SLOT.SIGNATURE)
         return x509.public_bytes(encoding=cert_format)
 
 
 @raise_yubikey_err("Cannot export public key.")
-def export_piv_pub_key(pub_key_format=serialization.Encoding.PEM, pub_key_pem=None):
+def export_piv_pub_key(pub_key_format=serialization.Encoding.PEM):
     """Exports YubiKey's piv slot public key.
 
     Args:
@@ -230,7 +207,7 @@ def export_piv_pub_key(pub_key_format=serialization.Encoding.PEM, pub_key_pem=No
     Raises:
         - YubikeyError
     """
-    with _yk_piv_ctrl(pub_key_pem=pub_key_pem) as (ctrl, _):
+    with _yk_piv_ctrl() as (ctrl, _):
         try:
             x509_cert = ctrl.get_certificate(SLOT.SIGNATURE)
             public_key = x509_cert.public_key()
@@ -256,7 +233,7 @@ def export_yk_certificate(certs_dir, key):
 
 
 @raise_yubikey_err("Cannot get public key in TUF format.")
-def get_piv_public_key_tuf(scheme=DEFAULT_RSA_SIGNATURE_SCHEME, pub_key_pem=None):
+def get_piv_public_key_tuf(scheme=DEFAULT_RSA_SIGNATURE_SCHEME):
     """Return public key from a Yubikey in TUF's RSAKEY_SCHEMA format.
 
     Args:
@@ -272,12 +249,12 @@ def get_piv_public_key_tuf(scheme=DEFAULT_RSA_SIGNATURE_SCHEME, pub_key_pem=None
     Raises:
         - YubikeyError
     """
-    pub_key_pem = export_piv_pub_key(pub_key_pem=pub_key_pem).decode("utf-8")
+    pub_key_pem = export_piv_pub_key().decode("utf-8")
     return import_rsakey_from_pem(pub_key_pem, scheme)
 
 
 @raise_yubikey_err("Cannot sign data.")
-def sign_piv_rsa_pkcs1v15(data, pin, pub_key_pem=None):
+def sign_piv_rsa_pkcs1v15(data, pin):
     """Sign data with key from YubiKey's piv slot.
 
     Args:
@@ -292,7 +269,7 @@ def sign_piv_rsa_pkcs1v15(data, pin, pub_key_pem=None):
     Raises:
         - YubikeyError
     """
-    with _yk_piv_ctrl(pub_key_pem=pub_key_pem) as (ctrl, _):
+    with _yk_piv_ctrl() as (ctrl, _):
         ctrl.verify_pin(pin)
         return ctrl.sign(
             SLOT.SIGNATURE, KEY_TYPE.RSA2048, data, hashes.SHA256(), padding.PKCS1v15()