+++ CHANGELOG.md
@@ -9,15 +9,15 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Added
 
+- Remove unused optional parameter from _yk_piv_ctrl ([572])
 - Implement full partial update. Store last validated commit per repo ([559)])
 
 ### Changed
 
 ### Fixed
 
-
+[572]: https://github.com/openlawlibrary/taf/pull/572
 [559]: https://github.com/openlawlibrary/taf/pull/558
-
 
 ## [0.32.4]
 
@@ -52,7 +52,6 @@ and this project adheres to [Semantic Versioning][semver].
 - Make url optional for `get_last_remote_commit` ([564])
 
 ### Fixed
-
 
 [564]: https://github.com/openlawlibrary/taf/pull/564